Replace print logs with zerolog in DynamoDB client
This commit replaces the print statements in DynamoDB client with zerolog for improved logging. Additional indirect dependencies are added in go.mod and go.sum files. Zerolog is now passed in the New function in dynamo.go file and used for logging in main.go, adding more structured and level-based logs.