Replace Redis with RabbitMQ and add RabbitMQ client
This commit swaps Redis with RabbitMQ for message queuing in the Docker Compose setup, which required the addition of new dependency libraries such as the RabbitMQ client and UUID library in the go.mod and go.sum files. A new RabbitMQ client was implemented in the `mq` package. The client setup is also reflected in changes to the Docker Compose file.