Replace ampq with mq package and implement RabbitMQ client
In the 'main' package, the ampq package was replaced with a newly implemented mq package to handle RabbitMQ connections instead of Redis. Additionally, the 'docker-compose.yml' file was amended to reflect these changes, including adding a health check for the RabbitMQ service, specifying hostnames, and providing necessary user credentials. These modifications enhance the efficiency and reliability of the service.